Big refactor! Player ids now rely on match ID instead of network ID.
When receiving a networkID, immediately translate it to a playerID through the TryGet(), and then use the ushort matchID from there on.

+++ Assets/Scripts/Game/Common/Phases/PhaseData/PhaseSyncLoadout.cs
@@ -12,7 +12,7 @@ namespace Game.Common.Phases.PhaseData
         public struct Data
         {
             [SerializeField] public bool valid;
-            [SerializeField] public uint[] PlayerIDs;
+            [SerializeField] public ushort[] PlayerIDs;
             [SerializeField] public int[] PlayerSelections;
         }
